Fiturr chatbot/requirements.txt
google-cloud-aiplatform >= 1.31.0
flask
google-cloud-logging
fastapi
pydantic
uvicorn
firebase-admin